data-sourcing/requirements.txt
httpx>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
anthropic>=0.39.0
pytrends>=4.9.0
